Refactor TwoFactorAuthenticatorsController#new
